// cmportal-frontend/src/types/TUser.ts
export type TUser = {
    appid: string;
    userid: string;
    firstName: string;
    lastName: string;
    emailAddress: string;
    isAdmin: string;
    functie: string;
    userToken: string;
    phoneNr: string;
    companyName: string;
    companyCode: string;
    cif: string;
    lang: string;
}