// cmportal-frontend/src/types/TAdresaLivrare.ts
export type TAdresaLivrare = {
    slid: string;
    appid: string;
    cif: string;
    adresaCodJudet: string;
    adresaLocalitate: string;
    adresaAdresa: string;
    adresaCodPostal: string;
    tipAdresa: string;
    codTara: string;
    navisionid: string;
    denJudet: string;
    isNewAddress: string;
}